
\name{create_books}
\alias{create_books}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Creates an empty ledgr database}
\description{
  Creates an empty ledger database
}
\usage{
create_books(path)
}
\arguments{
    \item{path}{the path you wish to contain the ledgr database}
}
\details{
    Blah
}
%- maybe also 'usage' for other objects documented here.
\references{}
\author{Bret Beheim}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }