% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_adapters.R
\name{ledgr_adapter_r}
\alias{ledgr_adapter_r}
\title{Adapt an R package function into a ledgr indicator}
\usage{
ledgr_adapter_r(pkg_fn, id, requires_bars, ...)
}
\arguments{
\item{pkg_fn}{R function (or "pkg::fn" string) to adapt.}

\item{id}{Indicator identifier.}

\item{requires_bars}{Minimum lookback period.}

\item{...}{Additional arguments passed to the package function.}
}
\value{
A \code{ledgr_indicator} object.
}
\description{
Adapt an R package function into a ledgr indicator
}
