% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{ledgr_compute_equity_curve}
\alias{ledgr_compute_equity_curve}
\title{Compute an equity curve from a backtest}
\usage{
ledgr_compute_equity_curve(bt)
}
\arguments{
\item{bt}{A \code{ledgr_backtest} object.}
}
\value{
A tibble containing equity, running maximum, and drawdown.
}
\description{
Compute an equity curve from a backtest
}
