% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ledgr_get_indicator}
\alias{ledgr_get_indicator}
\title{Get an indicator by name}
\usage{
ledgr_get_indicator(name)
}
\arguments{
\item{name}{Indicator name.}
}
\value{
A \code{ledgr_indicator} object.
}
\description{
Get an indicator by name
}
\examples{
local({
  registry <- get(".ledgr_indicator_registry", asNamespace("ledgr"))
  if (exists("example_lookup", envir = registry, inherits = FALSE)) {
    rm(list = "example_lookup", envir = registry)
  }
  on.exit(if (exists("example_lookup", envir = registry, inherits = FALSE)) {
    rm(list = "example_lookup", envir = registry)
  }, add = TRUE)
  ind <- ledgr_indicator(
    id = "example_lookup",
    fn = function(window) tail(window$close, 1),
    requires_bars = 1
  )
  ledgr_register_indicator(ind)
  ledgr_get_indicator("example_lookup")$id
})
}
