% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_builtin.R
\name{ledgr_ind_returns}
\alias{ledgr_ind_returns}
\title{Simple Returns}
\usage{
ledgr_ind_returns(n = 1L)
}
\arguments{
\item{n}{Periods back (default 1).}
}
\value{
A \code{ledgr_indicator} object.
}
\description{
Simple Returns
}
