% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_builtin.R
\name{ledgr_ind_rsi}
\alias{ledgr_ind_rsi}
\title{Relative Strength Index}
\usage{
ledgr_ind_rsi(n = 14L)
}
\arguments{
\item{n}{Window size (default 14).}
}
\value{
A \code{ledgr_indicator} object.
}
\description{
Relative Strength Index
}
\examples{
window <- data.frame(close = c(100, 101, 102, 101))
ledgr_ind_rsi(3)$fn(window)
}
