% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ledgr_list_indicators}
\alias{ledgr_list_indicators}
\title{List registered indicators}
\usage{
ledgr_list_indicators(pattern = NULL)
}
\arguments{
\item{pattern}{Optional regex filter for names.}
}
\value{
Character vector of indicator names.
}
\description{
List registered indicators
}
