% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ledgr_backtest}
\alias{plot.ledgr_backtest}
\title{Plot backtest results}
\usage{
\method{plot}{ledgr_backtest}(x, y = NULL, ..., type = "equity")
}
\arguments{
\item{x}{A \code{ledgr_backtest} object.}

\item{y}{Unused.}

\item{...}{Unused.}

\item{type}{Plot type. Only \code{"equity"} is supported in v0.1.2.}
}
\value{
A ggplot object, or a grid object when \code{gridExtra} is installed.
}
\description{
Plot backtest results
}
