% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{print.ledgr_experiment}
\alias{print.ledgr_experiment}
\title{Print a ledgr experiment}
\usage{
\method{print}{ledgr_experiment}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_experiment} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a ledgr experiment
}
