% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy-extract.R
\name{print.ledgr_extracted_strategy}
\alias{print.ledgr_extracted_strategy}
\title{Print extracted strategy metadata}
\usage{
\method{print}{ledgr_extracted_strategy}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_extracted_strategy} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print extracted strategy metadata
}
