% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{print.ledgr_indicator}
\alias{print.ledgr_indicator}
\title{Print a ledgr indicator}
\usage{
\method{print}{ledgr_indicator}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_indicator} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a ledgr indicator
}
\examples{
ledgr_ind_sma(20)
}
