% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-bundle.R
\name{print.ledgr_indicator_bundle}
\alias{print.ledgr_indicator_bundle}
\title{Print an indicator bundle}
\usage{
\method{print}{ledgr_indicator_bundle}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_indicator_bundle} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print an indicator bundle
}
\keyword{internal}
