% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_dev.R
\name{print.ledgr_indicator_dev}
\alias{print.ledgr_indicator_dev}
\title{Print an indicator development session}
\usage{
\method{print}{ledgr_indicator_dev}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_indicator_dev} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print an indicator development session
}
