% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{print.ledgr_opening}
\alias{print.ledgr_opening}
\title{Print an opening portfolio state}
\usage{
\method{print}{ledgr_opening}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_opening} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print an opening portfolio state
}
