% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-store.R
\name{print.ledgr_run_info}
\alias{print.ledgr_run_info}
\title{Print run information}
\usage{
\method{print}{ledgr_run_info}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_run_info} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print run information
}
