% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{print.ledgr_sweep_candidate}
\alias{print.ledgr_sweep_candidate}
\title{Print a sweep candidate}
\usage{
\method{print}{ledgr_sweep_candidate}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_sweep_candidate} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a sweep candidate
}
