% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{print.ledgr_sweep_results}
\alias{print.ledgr_sweep_results}
\title{Print sweep results}
\usage{
\method{print}{ledgr_sweep_results}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_sweep_results} object.}

\item{...}{Passed to the tibble print method.}
}
\value{
The input object, invisibly.
}
\description{
Print sweep results
}
