% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy-helpers.R
\name{weight_equal}
\alias{weight_equal}
\title{Convert a selection to equal long-only weights}
\usage{
weight_equal(selection)
}
\arguments{
\item{selection}{A \code{ledgr_selection} object.}
}
\value{
A \code{ledgr_weights} object.
}
\description{
\code{weight_equal()} assigns equal positive weights to selected instruments.
Empty selections produce empty weights.
}
\section{Articles}{

Strategy helper pipelines:
\code{vignette("strategy-development", package = "ledgr")}
\code{system.file("doc", "strategy-development.html", package = "ledgr")}
}

\examples{
selection <- ledgr_selection(c(AAA = TRUE, BBB = TRUE), universe = c("AAA", "BBB"))
weight_equal(selection)

}
