% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_dev.R
\name{close.ledgr_indicator_dev}
\alias{close.ledgr_indicator_dev}
\title{Close indicator development session}
\usage{
\method{close}{ledgr_indicator_dev}(con, ...)
}
\arguments{
\item{con}{A \code{ledgr_indicator_dev} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Close indicator development session
}
