% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_dev.R
\name{close.ledgr_pulse_context}
\alias{close.ledgr_pulse_context}
\title{Close pulse context}
\usage{
\method{close}{ledgr_pulse_context}(con, ...)
}
\arguments{
\item{con}{A \code{ledgr_pulse_context} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Close pulse context
}
