% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledgr-package.R
\docType{package}
\name{ledgr-package}
\alias{ledgr}
\alias{ledgr-package}
\title{ledgr: Event-Sourced Trading Framework Skeleton}
\description{
Correctness-first, event-sourced scaffolding for a trading lifecycle.
This package currently contains structure only; see
\code{inst/design/ledgr_design_document.md} for the binding specification.
}
\author{
\strong{Maintainer}: Max Thomasberger \email{31213680+blechturm@users.noreply.github.com}

}
\keyword{internal}
