% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ledgr_get_indicator}
\alias{ledgr_get_indicator}
\title{Get an indicator by name}
\usage{
ledgr_get_indicator(name)
}
\arguments{
\item{name}{Indicator name.}
}
\value{
A \code{ledgr_indicator} object.
}
\description{
Get an indicator by name
}
