% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_builtin.R
\name{ledgr_ind_ema}
\alias{ledgr_ind_ema}
\title{Exponential Moving Average}
\usage{
ledgr_ind_ema(n)
}
\arguments{
\item{n}{Window size.}
}
\value{
A \code{ledgr_indicator} object.
}
\description{
Exponential Moving Average
}
