% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_builtin.R
\name{ledgr_ind_sma}
\alias{ledgr_ind_sma}
\title{Simple Moving Average}
\usage{
ledgr_ind_sma(n)
}
\arguments{
\item{n}{Window size.}
}
\value{
A \code{ledgr_indicator} object.
}
\description{
Simple Moving Average
}
