% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ledgr_indicator}
\alias{ledgr_indicator}
\title{Construct a ledgr indicator}
\usage{
ledgr_indicator(
  id,
  fn,
  requires_bars,
  params = list(),
  stable_after = requires_bars
)
}
\arguments{
\item{id}{Unique indicator identifier.}

\item{fn}{Indicator function: function(window) -> numeric | list.}

\item{requires_bars}{Minimum lookback period (integer).}

\item{params}{Named list of deterministic parameters for fingerprinting.}

\item{stable_after}{Number of bars after which the indicator output is stable.}
}
\value{
A \code{ledgr_indicator} object.
}
\description{
Construct a ledgr indicator
}
