% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ledgr_register_indicator}
\alias{ledgr_register_indicator}
\title{Register an indicator in the global registry}
\usage{
ledgr_register_indicator(indicator, name = NULL)
}
\arguments{
\item{indicator}{A \code{ledgr_indicator} object.}

\item{name}{Optional registry name (defaults to indicator id).}
}
\value{
The indicator, invisibly.
}
\description{
Register an indicator in the global registry
}
