% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute-features.R
\name{print.ledgr_precomputed_features}
\alias{print.ledgr_precomputed_features}
\title{Print a precomputed feature payload}
\usage{
\method{print}{ledgr_precomputed_features}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_precomputed_features} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a precomputed feature payload
}
