% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_dev.R
\name{print.ledgr_pulse_context}
\alias{print.ledgr_pulse_context}
\title{Print a pulse snapshot context}
\usage{
\method{print}{ledgr_pulse_context}(x, ...)
}
\arguments{
\item{x}{A \code{ledgr_pulse_context} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a pulse snapshot context
}
